moved component in widgets
