+++ ui/src/views/compute/MigrateVMStorage.vue
@@ -46,7 +46,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'MigrateVMStorage',
