+++ ui/src/views/compute/StartVirtualMachine.vue
@@ -100,7 +100,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'StartVirtualMachine',
