+++ ui/src/views/compute/UpgradeKubernetesCluster.vue
@@ -58,7 +58,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'UpgradeKubernetesCluster',
