+++ ui/src/views/iam/ChangeUserPassword.vue
@@ -67,7 +67,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'ChangeUserPassword',
