+++ ui/src/views/iam/DomainActionForm.vue
@@ -131,7 +131,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'DomainActionForm',
