+++ ui/src/views/infra/network/EditTrafficLabel.vue
@@ -85,7 +85,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'EditTrafficLabel',
