+++ ui/src/views/network/CreateVlanIpRange.vue
@@ -155,7 +155,7 @@
 
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'CreateVlanIpRange',
