+++ ui/src/views/network/CreateVpc.vue
@@ -91,7 +91,7 @@
 </template>
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'CreateVpc',
