+++ ui/src/views/network/CreateVpnCustomerGateway.vue
@@ -231,7 +231,7 @@
 </template>
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'CreateVpnCustomerGateway',
