+++ ui/src/views/project/AddAccountOrUserToProject.vue
@@ -119,7 +119,7 @@
 </template>
 <script>
 import { api } from '@/api'
-import TooltipLabel from '@/components/view/TooltipLabel'
+import TooltipLabel from '@/components/widgets/TooltipLabel'
 
 export default {
   name: 'AddAccountOrUserToProject',
